Public runtime APIs for mapped plural task output properties
We’ve supported for some time `@OutputFiles` and `@OutputDirectories` properties with a `Map` type, where each output file or directory was mapped to a string identifier. However, this was only available via internal APIs and wasn’t documented. This commit exposes the runtime API via `TaskOutputs.files()` and the new `TaskOutputs.dirs()` methods.

+review REVIEW-6374